+++ tests/PolygonTest.php
@@ -1,0 +1,616 @@
+<?php
+
+declare(strict_types=1);
+
+namespace UIAwesome\Html\Svg\Tests;
+
+use InvalidArgumentException;
+use PHPUnit\Framework\Attributes\Group;
+use PHPUnit\Framework\TestCase;
+use UIAwesome\Html\Attribute\Values\{Aria, Data, Language, Role};
+use UIAwesome\Html\Core\Factory\SimpleFactory;
+use UIAwesome\Html\Helper\Enum;
+use UIAwesome\Html\Helper\Exception\Message;
+use UIAwesome\Html\Svg\Polygon;
+use UIAwesome\Html\Svg\Tests\Support\Stub\DefaultProvider;
+use UIAwesome\Html\Svg\Tests\Support\TestSupport;
+use UIAwesome\Html\Svg\Values\{FillRule, StrokeLineCap, StrokeLineJoin, SvgAttribute};
+
+/**
+ * Test suite for {@see Polygon} element functionality and behavior.
+ *
+ * Validates the management and rendering of the SVG `<polygon>` element according to the SVG 2 and HTML Living
+ * Standard specifications.
+ *
+ * Ensures correct handling, immutability, and validation of the `Polygon` tag rendering, supporting all global HTML,
+ * SVG 2 attributes and provider-based configuration.
+ *
+ * Test coverage.
+ * - Accurate rendering of the `<polygon>` element.
+ * - Correct application of global HTML attributes and SVG-specific attributes.
+ * - Error handling for invalid attribute values.
+ * - Immutability of the API, ensuring that setting attributes returns a new instance.
+ * - Integration with configuration providers and global factory defaults.
+ * - Precedence of user-defined attributes over global defaults and provider settings.
+ *
+ * {@see Polygon} for element implementation details.
+ * {@see SimpleFactory} for default configuration management.
+ * {@see TestSupport} for assertion utilities.
+ *
+ * @copyright Copyright (C) 2026 Terabytesoftw.
+ * @license https://opensource.org/license/bsd-3-clause BSD 3-Clause License.
+ */
+#[Group('svg')]
+final class PolygonTest extends TestCase
+{
+    use TestSupport;
+
+    public function testRenderWithAddAriaAttribute(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon aria-pressed="true">
+            HTML,
+            Polygon::tag()->addAriaAttribute('pressed', true)->render(),
+            "Failed asserting that element renders correctly with 'addAriaAttribute()' method.",
+        );
+    }
+
+    public function testRenderWithAddAriaAttributeUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon aria-pressed="true">
+            HTML,
+            Polygon::tag()->addAriaAttribute(Aria::PRESSED, true)->render(),
+            "Failed asserting that element renders correctly with 'addAriaAttribute()' method.",
+        );
+    }
+
+    public function testRenderWithAddDataAttribute(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon data-value="value">
+            HTML,
+            Polygon::tag()->addDataAttribute('value', 'value')->render(),
+            "Failed asserting that element renders correctly with 'addDataAttribute()' method.",
+        );
+    }
+
+    public function testRenderWithAddDataAttributeUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon data-value="value">
+            HTML,
+            Polygon::tag()->addDataAttribute(Data::VALUE, 'value')->render(),
+            "Failed asserting that element renders correctly with 'addDataAttribute()' method.",
+        );
+    }
+
+    public function testRenderWithAriaAttributes(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon aria-controls="modal-1" aria-hidden="false" aria-label="Close">
+            HTML,
+            Polygon::tag()
+                ->ariaAttributes(
+                    [
+                        'controls' => static fn(): string => 'modal-1',
+                        'hidden' => false,
+                        'label' => 'Close',
+                    ],
+                )
+                ->render(),
+            "Failed asserting that element renders correctly with 'ariaAttributes()' method.",
+        );
+    }
+
+    public function testRenderWithAttributes(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon class="value">
+            HTML,
+            Polygon::tag()->attributes(['class' => 'value'])->render(),
+            "Failed asserting that element renders correctly with 'attributes()' method.",
+        );
+    }
+
+    public function testRenderWithClass(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon class="value">
+            HTML,
+            Polygon::tag()->class('value')->render(),
+            "Failed asserting that element renders correctly with 'class' attribute.",
+        );
+    }
+
+    public function testRenderWithDataAttributes(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon data-value="test-value">
+            HTML,
+            Polygon::tag()->dataAttributes(['value' => 'test-value'])->render(),
+            "Failed asserting that element renders correctly with 'dataAttributes()' method.",
+        );
+    }
+
+    public function testRenderWithDefaultConfigurationValues(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon class="default-class">
+            HTML,
+            Polygon::tag(['class' => 'default-class'])->render(),
+            'Failed asserting that default configuration values are applied correctly.',
+        );
+    }
+
+    public function testRenderWithDefaultProvider(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon class="default-class" title="default-title">
+            HTML,
+            Polygon::tag()->addDefaultProvider(DefaultProvider::class)->render(),
+            'Failed asserting that default provider is applied correctly.',
+        );
+    }
+
+    public function testRenderWithFill(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon fill="#ff0000">
+            HTML,
+            Polygon::tag()->fill('#ff0000')->render(),
+            "Failed asserting that element renders correctly with 'fill' attribute.",
+        );
+    }
+
+    public function testRenderWithFillOpacity(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon fill-opacity="0.7">
+            HTML,
+            Polygon::tag()->fillOpacity('0.7')->render(),
+            "Failed asserting that element renders correctly with 'fill-opacity' attribute.",
+        );
+    }
+
+    public function testRenderWithFillRule(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon fill-rule="evenodd">
+            HTML,
+            Polygon::tag()->fillRule('evenodd')->render(),
+            "Failed asserting that element renders correctly with 'fill-rule' attribute.",
+        );
+    }
+
+    public function testRenderWithFillRuleUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon fill-rule="nonzero">
+            HTML,
+            Polygon::tag()->fillRule(FillRule::NONZERO)->render(),
+            "Failed asserting that element renders correctly with 'fill-rule' attribute.",
+        );
+    }
+
+    public function testRenderWithGlobalDefaultsAreApplied(): void
+    {
+        SimpleFactory::setDefaults(Polygon::class, ['class' => 'default-class']);
+
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon class="default-class">
+            HTML,
+            Polygon::tag()->render(),
+            'Failed asserting that global defaults are applied correctly.',
+        );
+
+        SimpleFactory::setDefaults(Polygon::class, []);
+    }
+
+    public function testRenderWithId(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon id="test-id">
+            HTML,
+            Polygon::tag()->id('test-id')->render(),
+            "Failed asserting that element renders correctly with 'id' attribute.",
+        );
+    }
+
+    public function testRenderWithLangUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon lang="es">
+            HTML,
+            Polygon::tag()->lang(Language::SPANISH)->render(),
+            "Failed asserting that element renders correctly with 'lang' attribute.",
+        );
+    }
+
+    public function testRenderWithOpacity(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon opacity="0.5">
+            HTML,
+            Polygon::tag()->opacity('0.5')->render(),
+            "Failed asserting that element renders correctly with 'opacity' attribute.",
+        );
+    }
+
+    public function testRenderWithPathLength(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon pathLength="200">
+            HTML,
+            Polygon::tag()->pathLength(200)->render(),
+            "Failed asserting that element renders correctly with 'pathLength' attribute.",
+        );
+    }
+
+    public function testRenderWithPoints(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon points="0,100 50,25 50,75 100,0">
+            HTML,
+            Polygon::tag()->points('0,100 50,25 50,75 100,0')->render(),
+            "Failed asserting that element renders correctly with 'points' attribute.",
+        );
+    }
+
+    public function testRenderWithRole(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon role="banner">
+            HTML,
+            Polygon::tag()->role('banner')->render(),
+            "Failed asserting that element renders correctly with 'role' attribute.",
+        );
+    }
+
+    public function testRenderWithRoleUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon role="banner">
+            HTML,
+            Polygon::tag()->role(Role::BANNER)->render(),
+            "Failed asserting that element renders correctly with 'role' attribute.",
+        );
+    }
+
+    public function testRenderWithStroke(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon stroke="#00ff00">
+            HTML,
+            Polygon::tag()->stroke('#00ff00')->render(),
+            "Failed asserting that element renders correctly with 'stroke' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeDashArray(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon stroke-dasharray="5,5">
+            HTML,
+            Polygon::tag()->strokeDashArray('5,5')->render(),
+            "Failed asserting that element renders correctly with 'stroke-dasharray' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeLineCap(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon stroke-linecap="round">
+            HTML,
+            Polygon::tag()->strokeLineCap('round')->render(),
+            "Failed asserting that element renders correctly with 'stroke-linecap' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeLineCapUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon stroke-linecap="square">
+            HTML,
+            Polygon::tag()->strokeLineCap(StrokeLineCap::SQUARE)->render(),
+            "Failed asserting that element renders correctly with 'stroke-linecap' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeLineJoin(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon stroke-linejoin="round">
+            HTML,
+            Polygon::tag()->strokeLineJoin('round')->render(),
+            "Failed asserting that element renders correctly with 'stroke-linejoin' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeLineJoinUsingEnum(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon stroke-linejoin="round">
+            HTML,
+            Polygon::tag()->strokeLineJoin(StrokeLineJoin::ROUND)->render(),
+            "Failed asserting that element renders correctly with 'stroke-linejoin' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeMiterlimit(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon stroke-miterlimit="10">
+            HTML,
+            Polygon::tag()->strokeMiterlimit(10)->render(),
+            "Failed asserting that element renders correctly with 'stroke-miterlimit' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeOpacity(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon stroke-opacity="0.8">
+            HTML,
+            Polygon::tag()->strokeOpacity(0.8)->render(),
+            "Failed asserting that element renders correctly with 'stroke-opacity' attribute.",
+        );
+    }
+
+    public function testRenderWithStrokeWidth(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon stroke-width="2">
+            HTML,
+            Polygon::tag()->strokeWidth('2')->render(),
+            "Failed asserting that element renders correctly with 'stroke-width' attribute.",
+        );
+    }
+
+    public function testRenderWithStyle(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon style='test-value'>
+            HTML,
+            Polygon::tag()->style('test-value')->render(),
+            "Failed asserting that element renders correctly with 'style' attribute.",
+        );
+    }
+
+    public function testRenderWithToString(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon>
+            HTML,
+            (string) Polygon::tag(),
+            "Failed asserting that '__toString()' method renders correctly.",
+        );
+    }
+
+    public function testRenderWithTransform(): void
+    {
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon transform="rotate(45)">
+            HTML,
+            Polygon::tag()->transform('rotate(45)')->render(),
+            "Failed asserting that element renders correctly with 'transform' attribute.",
+        );
+    }
+
+    public function testRenderWithUserOverridesGlobalDefaults(): void
+    {
+        SimpleFactory::setDefaults(Polygon::class, ['class' => 'from-global', 'id' => 'id-global']);
+
+        self::equalsWithoutLE(
+            <<<HTML
+            <polygon class="from-global" id="id-user">
+            HTML,
+            Polygon::tag(['id' => 'id-user'])->render(),
+            'Failed asserting that user-defined attributes override global defaults correctly.',
+        );
+
+        SimpleFactory::setDefaults(Polygon::class, []);
+    }
+
+    public function testReturnNewInstanceWhenSettingAttribute(): void
+    {
+        $polygon = Polygon::tag();
+
+        self::assertNotSame(
+            $polygon,
+            $polygon->fill(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $polygon,
+            $polygon->fillOpacity('0'),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $polygon,
+            $polygon->fillRule(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $polygon,
+            $polygon->opacity('0'),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $polygon,
+            $polygon->pathLength(0),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $polygon,
+            $polygon->points(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $polygon,
+            $polygon->stroke(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $polygon,
+            $polygon->strokeDashArray(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $polygon,
+            $polygon->strokeLineCap(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $polygon,
+            $polygon->strokeLineJoin(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $polygon,
+            $polygon->strokeMiterlimit('1'),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $polygon,
+            $polygon->strokeOpacity('0'),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $polygon,
+            $polygon->strokeWidth(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+        self::assertNotSame(
+            $polygon,
+            $polygon->transform(''),
+            'Should return a new instance when setting the attribute, ensuring immutability.',
+        );
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingFillOpacityValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            \UIAwesome\Html\Svg\Exception\Message::VALUE_OUT_OF_RANGE_OR_NULL->getMessage(0, 1),
+        );
+
+        Polygon::tag()->fillOpacity('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingFillRuleValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            Message::VALUE_NOT_IN_LIST->getMessage(
+                'invalid-value',
+                SvgAttribute::FILL_RULE->value,
+                implode('\', \'', Enum::normalizeArray(FillRule::cases())),
+            ),
+        );
+
+        Polygon::tag()->fillRule('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingOpacityValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            \UIAwesome\Html\Svg\Exception\Message::VALUE_OUT_OF_RANGE_OR_NULL->getMessage(0, 1),
+        );
+
+        Polygon::tag()->opacity('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingPathLengthValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            \UIAwesome\Html\Svg\Exception\Message::VALUE_MUST_BE_POSITIVE_NUMBER_OR_NULL->getMessage(),
+        );
+
+        Polygon::tag()->pathLength('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingStrokeLineCapValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            Message::VALUE_NOT_IN_LIST->getMessage(
+                'invalid-value',
+                SvgAttribute::STROKE_LINECAP->value,
+                implode('\', \'', Enum::normalizeArray(StrokeLineCap::cases())),
+            ),
+        );
+
+        Polygon::tag()->strokeLineCap('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingStrokeLineJoinValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            Message::VALUE_NOT_IN_LIST->getMessage(
+                'invalid-value',
+                SvgAttribute::STROKE_LINEJOIN->value,
+                implode('\', \'', Enum::normalizeArray(StrokeLineJoin::cases())),
+            ),
+        );
+
+        Polygon::tag()->strokeLineJoin('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingStrokeMiterlimitValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            \UIAwesome\Html\Svg\Exception\Message::VALUE_MUST_BE_GTE_ONE_OR_NULL->getMessage(),
+        );
+
+        Polygon::tag()->strokeMiterlimit('invalid-value');
+    }
+
+    public function testThrowInvalidArgumentExceptionForSettingStrokeOpacityValue(): void
+    {
+        $this->expectException(InvalidArgumentException::class);
+        $this->expectExceptionMessage(
+            \UIAwesome\Html\Svg\Exception\Message::VALUE_OUT_OF_RANGE_OR_NULL->getMessage(0, 1),
+        );
+
+        Polygon::tag()->strokeOpacity('invalid-value');
+    }
+}